Remove arquivos de testes de dados seprados e integra aos testes existentes
